+++ .storybook/addons.js
@@ -4,3 +4,4 @@ import "@storybook/addon-storysource/register";
 import "@storybook/addon-viewport/register";
 import "storybook-styled-components/register";
 import "@storybook/addon-actions/register";
+import "@storybook/addon-docs/register";