+++ src/@next/components/atoms/IconButton/stories.tsx
@@ -4,5 +4,6 @@ import React from "react";
 import { IconButton } from ".";
 
 storiesOf("@components/atoms/IconButton", module)
+  .addParameters({ component: IconButton })
   .add("edit icon button", () => <IconButton name="edit" size={19} />)
   .add("trash icon button", () => <IconButton name="trash" size={22} />);