+++ src/@next/components/atoms/NavLink/stories.tsx
@@ -5,8 +5,10 @@ import { BrowserRouter } from "react-router-dom";
 import { NavLink } from ".";
 import { mockItemRoute } from "./fixtures";
 
-storiesOf("@components/atoms/NavLink", module).add("default", () => (
-  <BrowserRouter>
-    <NavLink item={mockItemRoute}>ELEMENT I</NavLink>
-  </BrowserRouter>
-));
+storiesOf("@components/atoms/NavLink", module)
+  .addParameters({ component: NavLink })
+  .add("default", () => (
+    <BrowserRouter>
+      <NavLink item={mockItemRoute}>ELEMENT I</NavLink>
+    </BrowserRouter>
+  ));