+++ src/@next/components/atoms/Tile/stories.tsx
@@ -4,6 +4,7 @@ import React from "react";
 import { Tile } from ".";
 
 storiesOf("@components/atoms/Tile", module)
+  .addParameters({ component: Tile })
   .add("default", () => (
     <Tile header={<h3>This is header</h3>}>
       <div>This is body</div>