+++ src/@next/components/molecules/OrderTabel/stories.tsx
@@ -73,9 +73,11 @@ const ORDERS = [
   },
 ];
 
-storiesOf("@components/molecules/OrderTabel", module).add("default", () => {
-  // this is added to fix issue when running on machines with different locales
-  // so one one machine price is displayed as $42.24 and on others as US$ 42.24 etc
-  Number.prototype.toLocaleString = () => "";
-  return <OrderTabel history={history} orders={ORDERS} />;
-});
+storiesOf("@components/molecules/OrderTabel", module)
+  .addParameters({ component: OrderTabel })
+  .add("default", () => {
+    // this is added to fix issue when running on machines with different locales
+    // so one one machine price is displayed as $42.24 and on others as US$ 42.24 etc
+    Number.prototype.toLocaleString = () => "";
+    return <OrderTabel history={history} orders={ORDERS} />;
+  });