+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix not working storefront when no data in saleor database exist - #551 by @orzechdev
 - Make checkout working without shipping if it is not required - #571 by @orzechdev
 - Add ability to apply a promo code in checkout - #582 by @orzechdev
+- Add docs to storybook - #614 by @orzechdev
 
 ## 0.7.0